requirements.txt
streamlit
opencv-python
pillow
pytesseract
openai
python-dotenv
fpdf
